Refactor TT
